jobsy-up-school/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2 